Pass task 2.4

+++ task04/src/com/example/task04/Task04.java
@@ -13,7 +13,7 @@ public class Task04 {
             case '*':
                 return a*b;
             case '/':
-                return a/b;
+                return (float)a/(float)b;
         }
 
         return 0;
